后端提交(2023/3/5) 1. 完善表的公共字段 2. 添加README.md 3. 添加：简单的用户查询功能 4. 调整接口：笔记项目列举、笔记项目详细信息

+++ src/main/java/org/sand/model/po/note/NotePO.java
@@ -10,7 +10,7 @@ import lombok.Data;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class NotePO {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -21,6 +21,6 @@ public class NotePO {
     private String path;
 
     @TableField("project_id")
-    private Integer projectId;
+    private Long projectId;
 
 }
